LPS-206206 Extract common logic to a base class, no logic changes
