+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/frontend/taglib/servlet/taglib/LayoutDesignScreenNavigationEntry.java
@@ -5,62 +5,27 @@
 
 package com.liferay.layout.admin.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
-import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategoryProvider;
-import com.liferay.frontend.taglib.form.navigator.FormNavigatorEntryProvider;
-import com.liferay.frontend.taglib.form.navigator.constants.FormNavigatorConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.layout.admin.constants.LayoutScreenNavigationEntryConstants;
 import com.liferay.layout.constants.LayoutTypeSettingsConstants;
-import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.util.AggregateResourceBundle;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 
-import java.io.IOException;
-
-import java.util.List;
 import java.util.Locale;
-import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Sandro Chinea
  */
 @Component(service = ScreenNavigationEntry.class)
 public class LayoutDesignScreenNavigationEntry
-	implements ScreenNavigationEntry<Layout> {
-
-	@Override
-	public String getCategoryKey() {
-		return LayoutScreenNavigationEntryConstants.CATEGORY_KEY_GENERAL;
-	}
+	extends BaseLayoutScreenNavigationEntry {
 
 	@Override
 	public String getEntryKey() {
 		return LayoutScreenNavigationEntryConstants.ENTRY_KEY_DESIGN;
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return _language.get(_getResourceBundle(locale), getEntryKey());
-	}
-
-	@Override
-	public String getScreenNavigationKey() {
-		return LayoutScreenNavigationEntryConstants.
-			SCREEN_NAVIGATION_KEY_LAYOUT;
 	}
 
 	@Override
@@ -83,65 +48,15 @@ public class LayoutDesignScreenNavigationEntry
 					LayoutTypeSettingsConstants.
 						KEY_DESIGN_CONFIGURATION_MODIFIED))) {
 
-			return _language.get(_getResourceBundle(locale), "draft");
+			return language.get(getResourceBundle(locale), "draft");
 		}
 
 		return null;
 	}
 
 	@Override
-	public boolean isVisible(User user, Layout layout) {
-		List<FormNavigatorCategory> formNavigatorCategories =
-			_formNavigatorCategoryProvider.getFormNavigatorCategories(
-				FormNavigatorConstants.FORM_NAVIGATOR_ID_LAYOUT_DESIGN);
-
-		for (FormNavigatorCategory formNavigatorCategory :
-				formNavigatorCategories) {
-
-			if (ListUtil.isNotEmpty(
-					_formNavigatorEntryProvider.getFormNavigatorEntries(
-						FormNavigatorConstants.FORM_NAVIGATOR_ID_LAYOUT_DESIGN,
-						formNavigatorCategory.getKey(), user, layout))) {
-
-				return true;
-			}
-		}
-
-		return false;
+	protected String getJspPath() {
+		return "/layout/screen/navigation/entries/design.jsp";
 	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/layout/screen/navigation/entries/design.jsp");
-	}
-
-	private ResourceBundle _getResourceBundle(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, getClass());
-
-		return new AggregateResourceBundle(
-			resourceBundle, _portal.getResourceBundle(locale));
-	}
-
-	@Reference
-	private FormNavigatorCategoryProvider _formNavigatorCategoryProvider;
-
-	@Reference
-	private FormNavigatorEntryProvider _formNavigatorEntryProvider;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private Portal _portal;
 
 }
